requirements.txt
pandas
numpy
faker
pyyaml
openpyxl
xlsxwriter
pytest
